+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.3 (wip)
+
+- Add `VTAcknowledgement` model class
+- Make `acknowledgements` property for `VTAcknowledgementsViewController` public
+
 ## 0.2 (27 December 2013)
 
 - Non-editable license text view (duh)
@@ -7,3 +12,5 @@
 - Customizable font for license text view
 
 ## 0.1 (24 December 2013)
+
+Initial import